refactor(analyze): extract types.ts and language-profiles.ts (#5)
Move all TypeScript interfaces and language configuration into standalone
modules, reducing index.ts coupling and enabling independent reuse.

- Create tools/analyze/types.ts — all interfaces (RefactorFlags, LanguageProfile,
  Tier1Result, Responsibility, SplitSuggestion, Tier2Result, AnalysisResult,
  CacheEntry, AnalysisCache, RefactorReport, ReportSummary, IssueData)
- Create tools/analyze/language-profiles.ts — LANGUAGE_PROFILES array,
  DEFAULT_PROFILE, and SOURCE_EXTENSIONS set
- Update index.ts to import from both new modules; no behavior change

No circular dependencies: types.ts has no project imports,
language-profiles.ts imports only from types.ts.
